refactor: Fix build breaking refactor changes
