+++ packages/cli/src/keys.ts
@@ -7,11 +7,7 @@ const c32check = require('c32check');
 import { HDKey } from '@scure/bip32';
 import * as scureBip39 from '@scure/bip39';
 
-import {
-  compressPrivateKey,
-  getPublicKeyFromPrivate,
-  publicKeyToBtcAddress,
-} from '@stacks/encryption';
+import { getPublicKeyFromPrivate, publicKeyToBtcAddress } from '@stacks/encryption';
 import { DerivationType, deriveAccount, generateWallet, getRootNode } from '@stacks/wallet-sdk';
 import * as bip32 from 'bip32';
 import * as bip39 from 'bip39';
@@ -20,6 +16,7 @@ import * as wif from 'wif';
 
 import { getMaxIDSearchIndex, getPrivateKeyAddress } from './common';
 import { CLINetworkAdapter } from './network';
+import { compressPrivateKey } from '@stacks/transactions';
 
 const BITCOIN_PUBKEYHASH = 0;
 const BITCOIN_PUBKEYHASH_TESTNET = 111;
